requirements.txt
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
flask-wtf==1.2.1
flask-bcrypt~=1.0.1
flask-testing~=0.8.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector-python==8.4.0
mock~=5.1.0
requests==2.31.0
urllib3==2.2.1
Werkzeug==3.0.3
WTForms~=3.1.2
python-dotenv>=0.19.2
authlib>=1.0
flask_sqlalchemy>=3.0.1
sqlalchemy
